RegisterUserTests added
